Improve Flex storage

+++ system/src/Grav/Framework/Flex/Storage/FolderStorage.php
@@ -64,7 +64,7 @@ class FolderStorage extends AbstractFilesystemStorage
      */
     public function hasKey(string $key) : bool
     {
-        return $key && file_exists($this->getPathFromKey($key));
+        return $key && !strpos($key, '@@') && file_exists($this->getPathFromKey($key));
     }
 
     /**
@@ -93,9 +93,13 @@ class FolderStorage extends AbstractFilesystemStorage
         foreach ($rows as $key => $row) {
             if (null === $row || (!\is_object($row) && !\is_array($row))) {
                 // Only load rows which haven't been loaded before.
-                $path = $this->getPathFromKey($key);
-                $file = $this->getFile($path);
-                $list[$key] = $this->hasKey($key) ? $this->loadFile($file) : null;
+                if (!$this->hasKey($key)) {
+                    $list[$key] = null;
+                } else {
+                    $path = $this->getPathFromKey($key);
+                    $file = $this->getFile($path);
+                    $list[$key] = $this->loadFile($file);
+                }
                 if (null !== $fetched) {
                     $fetched[$key] = $list[$key];
                 }
@@ -115,9 +119,13 @@ class FolderStorage extends AbstractFilesystemStorage
     {
         $list = [];
         foreach ($rows as $key => $row) {
-            $path = $this->getPathFromKey($key);
-            $file = $this->getFile($path);
-            $list[$key] = $this->hasKey($key) ? $this->saveFile($file, $row) : null;
+            if (!$this->hasKey($key)) {
+                $list[$key] = null;
+            } else {
+                $path = $this->getPathFromKey($key);
+                $file = $this->getFile($path);
+                $list[$key] = $this->saveFile($file, $row);
+            }
         }
 
         return $list;
@@ -130,15 +138,19 @@ class FolderStorage extends AbstractFilesystemStorage
     {
         $list = [];
         foreach ($rows as $key => $row) {
-            $path = $this->getPathFromKey($key);
-            $file = $this->getFile($path);
-            $list[$key] = $this->hasKey($key) ? $this->deleteFile($file) : null;
+            if (!$this->hasKey($key)) {
+                $list[$key] = null;
+            } else {
+                $path = $this->getPathFromKey($key);
+                $file = $this->getFile($path);
+                $list[$key] = $this->deleteFile($file);
 
-            $storage = $this->getStoragePath($key);
-            $media = $this->getMediaPath($key);
+                $storage = $this->getStoragePath($key);
+                $media = $this->getMediaPath($key);
 
-            $this->deleteFolder($storage, true);
-            $media && $this->deleteFolder($media, true);
+                $this->deleteFolder($storage, true);
+                $media && $this->deleteFolder($media, true);
+            }
         }
 
         return $list;
@@ -151,6 +163,9 @@ class FolderStorage extends AbstractFilesystemStorage
     {
         $list = [];
         foreach ($rows as $key => $row) {
+            if (strpos($key, '@@')) {
+                $key = $this->getNewKey();
+            }
             $path = $this->getPathFromKey($key);
             $file = $this->getFile($path);
             $list[$key] = $this->saveFile($file, $row);
